+++ src/components/Chain/ChainEcosystem/ChainEcosystem.tsx
@@ -4,11 +4,11 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { ItemV } from '../../../src/css/SharedStyling';
-import ChainBottomComponent from '../Chain/ChainBottomComponent';
+import { ItemV } from '../../../css/SharedStyling';
+import ChainBottomComponent from '../ChainBottomComponent';
 import ChainKnowledgeBaseGrid from '../ChainKnowledgeBase/ChainKnowledgeGrid';
 
-import { device } from '../../../src/config/globals';
+import { device } from '../../../config/globals';
 import ChainEcosystemContent from './ChainEcosystemContent';
 
 const ChainEcosystem = () => {
