+++ src/components/Chain/ChainKnowledgeBase/ChainKnowledgeBaseComponent.tsx
@@ -3,11 +3,11 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { device } from '../../config/globals';
-import useMediaQuery from '../../hooks/useMediaQuery';
+import { device } from '../../../config/globals';
+import useMediaQuery from '../../../hooks/useMediaQuery';
 
-import { Button, H2, H3, ItemH, ItemV } from '../../css/SharedStyling';
-import { ChainKnowledgeBaseHeaderConfig } from '../../config/ChainKnowledgeBaseConfig';
+import { Button, H2, H3, ItemH, ItemV } from '../../../css/SharedStyling';
+import { ChainKnowledgeBaseHeaderConfig } from '../../../config/ChainKnowledgeBaseConfig';
 
 const ChainKnowledgeBaseComponent = () => {
   const isMobile = useMediaQuery(device.mobileL);
