+++ src/components/Chain/ChainKnowledgeBase/ChannelKnowledgeBaseComponentItem.tsx
@@ -4,10 +4,10 @@ import React, { FC } from 'react';
 
 import { TbArrowUpRight } from 'react-icons/tb';
 
-import { device } from '../../config/globals';
-import useMediaQuery from '../../hooks/useMediaQuery';
+import { device } from '../../../config/globals';
+import useMediaQuery from '../../../hooks/useMediaQuery';
 
-import { H3, ItemH, ItemV, Span } from '../../css/SharedStyling';
+import { H3, ItemH, ItemV, Span } from '../../../css/SharedStyling';
 
 const ChannelKnowledgeBaseComponentItem: FC = ({ item, index }) => {
   const isMobile = useMediaQuery(device.mobileL);
