+++ src/components/Chain/ChainKnowledgeBaseArticle/ChainKnowledgeBaseArticle.tsx
@@ -4,18 +4,18 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { ItemV } from '../../../src/css/SharedStyling';
-import ChainBottomComponent from '../Chain/ChainBottomComponent';
-import ChainKnowledgeBaseComponent from './ChainKnowledgeBaseComponent';
-import ChainKnowledgeBaseGrid from './ChainKnowledgeGrid';
+import { ItemV } from '../../../css/SharedStyling';
+import ChainBottomComponent from '../ChainBottomComponent';
+import ChainKnowledgeBaseGrid from '../ChainKnowledgeBase/ChainKnowledgeGrid';
+import ChainKnowledgeBaseArticleContent from './ChainKnowledgeBaseArticleContent';
 
-import { device } from '../../../src/config/globals';
+import { device } from '../../../config/globals';
 
-const ChainKnowledgeBase = () => {
+const ChainKnowledgeBaseArticle = ({ item }) => {
   return (
     <ChainWrapper>
       <ChainInnerWrapper>
-        <ChainKnowledgeBaseComponent />
+        <ChainKnowledgeBaseArticleContent item={item} />
         <ChainKnowledgeBaseGrid />
         <ChainBottomComponent />
       </ChainInnerWrapper>
@@ -23,7 +23,7 @@ const ChainKnowledgeBase = () => {
   );
 };
 
-export default ChainKnowledgeBase;
+export default ChainKnowledgeBaseArticle;
 
 const ChainWrapper = styled(ItemV)`
   width: 100%;
