+++ src/components/Chain/ChainKnowledgeBaseArticle/ChainKnowledgeBaseArticleContent.tsx
@@ -5,64 +5,19 @@ import styled from 'styled-components';
 import Markdown from 'react-markdown';
 import rehypeRaw from 'rehype-raw';
 import remarkGfm from 'remark-gfm';
-import { unified } from 'unified';
-import remarkParse from 'remark-parse';
-import { visit } from 'unist-util-visit';
 import rehypeSlug from 'rehype-slug';
-import rehypeStringify from 'rehype-stringify';
-import remarkRehype from 'remark-rehype';
 
-import { device } from '../../config/globals';
-import useMediaQuery from '../../hooks/useMediaQuery';
+import { device } from '../../../config/globals';
+import useMediaQuery from '../../../hooks/useMediaQuery';
+import { extractTOC } from '../utils/ExtractTableOfContent';
 
-import { H3, ItemH, ItemV, Span } from '../../css/SharedStyling';
+import { H3, ItemH, ItemV, Span } from '../../../css/SharedStyling';
 import Link from '@docusaurus/Link';
 
 const ChainKnowledgeBaseArticleContent = ({ item }) => {
   const isMobile = useMediaQuery(device.mobileL);
   const isTablet = useMediaQuery(device.tablet);
   const { content } = item;
-
-  const extractTOC = (markdownContent, headingLevels = [1, 2, 3]) => {
-    const toc = [];
-
-    const processor = unified()
-      .use(remarkParse) // Parse Markdown
-      .use(remarkRehype) // Convert to HTML AST
-      .use(rehypeSlug) // Add slug IDs to headings
-      .use(rehypeStringify); // Convert back to HTML string (optional)
-
-    // Process the Markdown content through the pipeline
-    const tree = processor.runSync(processor.parse(markdownContent));
-
-    visit(tree, 'element', (node) => {
-      if (
-        node.tagName &&
-        ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'].includes(node.tagName) &&
-        headingLevels.includes(parseInt(node.tagName[1], 10))
-      ) {
-        const id = node.properties?.id || '';
-
-        // Extract the text content from the heading node
-        const text = node.children
-          .map((child) => {
-            if (child.type === 'text') return child.value; // Direct text node
-            if (child.children) {
-              // Nested children (e.g., emphasis or links)
-              return child.children
-                .map((nestedChild) => nestedChild.value || '')
-                .join('');
-            }
-            return '';
-          })
-          .join('');
-
-        toc.push({ level: parseInt(node.tagName[1], 10), text, id });
-      }
-    });
-
-    return toc;
-  };
 
   const toc = extractTOC(content);
 
@@ -153,7 +108,8 @@ const MarkdownSection = styled(ItemV)`
   ol {
     li {
       list-style: inherit !important;
-    }
+    }import { extractTOC } from '../utils/ExtractTableOfContent';
+
   }
 `;
